// resources/app/libs/api-client/gen/models/appHttpIntegrationsMusicBrainzModelsRelease.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Bånder api docs
 * Bånder.App is a sophisticated media server developed with the primary goal to deliver high performance and seamless experience for users.
Being built on top of Laravel framework and PostgreSQL for database, it leverages the power of robust backend technologies, providing high security, reliability, and scalability.

The main focus of Bånder.App is efficient media management and delivery.
This application is designed to handle large amounts of media files while ensuring quick and efficient access.
Implemented queuing functionality through Redis offers efficient job management and load handling.
The media information is organized and easily searchable, ensuring users can always find what they are looking for quickly and easily.
 * OpenAPI spec version: 0.0.1
 */

export interface AppHttpIntegrationsMusicBrainzModelsRelease {
  release?: number;
  id?: string;
  title?: string;
  status?: string;
  quality?: string;
  annotation?: string;
  text?: string;
  type?: string;
  entity?: string;
  name?: string;
  disambiguation?: string;
  packaging?: string;
  text_representation?: unknown[];
  language?: string;
  script?: string;
  artist_credit?: unknown[];
  count?: number;
  offset?: number;
  alias?: unknown[];
  locale?: string;
  sort_name?: string;
  type_id?: string;
  primary?: boolean;
  begin_date?: string;
  end_date?: string;
  release_group?: unknown[];
  first_release_date?: string;
  primary_type?: string;
  secondary_type?: unknown[];
  date?: string;
  country?: string;
  release_event?: unknown[];
  area?: unknown[];
  iso_3166_1_code?: string;
  iso_3166_2_code?: string;
  iso_3166_3_code?: string;
  life_span?: unknown[];
  begin?: string;
  end?: string;
  ended?: boolean;
  tag?: unknown[];
  user_tag?: unknown[];
  genre?: unknown[];
  target_type?: string;
  relation?: unknown[];
  target?: string;
  ordering_key?: number;
  direction?: string;
  source_credit?: string;
  target_credit?: string;
  user_genre?: unknown[];
  barcode?: string;
  asin?: string;
  cover_art_archive?: unknown[];
  artwork?: boolean;
  front?: boolean;
  back?: boolean;
  darkened?: boolean;
  label_info?: unknown[];
  catalog_number?: string;
  label?: unknown[];
  label_code?: string;
  ipi?: unknown[];
  isni?: unknown[];
  rating?: unknown[];
  votes_count?: number;
  user_rating?: number;
  track_count?: number;
  medium?: unknown[];
  track?: unknown[];
  position?: number;
  number?: string;
  length?: number;
  recording?: unknown[];
  video?: boolean;
  puid?: unknown[];
  isrc?: unknown[];
  format?: unknown[];
  disc?: unknown[];
  sectors?: number;
  pregap?: number;
  collection?: unknown[];
  entity_type?: string;
  editor?: string;
  artist?: unknown[];
  gender?: string;
  begin_area?: unknown[];
  end_area?: unknown[];
  work?: unknown[];
  fluency?: string;
  iswc?: unknown[];
  event?: unknown[];
  cancelled?: boolean;
  time?: string;
  setlist?: unknown[];
  instrument?: unknown[];
  description?: string;
  place?: unknown[];
  address?: string;
  coordinates?: unknown[];
  latitude?: number;
  longitude?: number;
  series?: unknown[];
  media?: unknown[];
  score?: number;
}